pipecat_requirements.txt
pipecat-ai[daily,deepgram,cartesia,openai,silero]>=0.0.68
python-dotenv>=1.0.0
aiohttp>=3.8.0 